# apis/api.yaml
openapi: 3.0.0
info:
  title: Canvara Backend API
  version: 0.11.0
servers: []
...
paths:
  /activities/my:
    get:
      summary: List my activities
      operationId: list_activities
      tags:
        - activities
      security:
        - AWSCognitoCanvaraPrototype: []
      parameters:
        - schema:
            type: integer
            default: 20
          in: query
          name: limit
          description: Maximum number of activities to return.
        - schema:
            type: integer
          in: query
          name: start
          description: Start index from where to return the activities.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  activities:
                    type: array
                    items:
                      type: object
                      properties:
                        activity_id:
                          type: string
                          description: Unique identifier of this activity.
                        type:
                          type: string
                          description: Type of activity.
                        data:
                          type: object
                          description: Data to display in the activity.
                        created_at:
                          type: string
                          description: 'UTC date and time when this activity was created, in ISO 8601 format.'
                        status:
                          type: string
                          description: Status of the activity.
                          enum:
                            - read
                            - unread
                  total_unread:
                    type: number
  /activities/global:
    get:
      summary: List global activities
      operationId: list_activities_global
      tags:
        - activities
      security:
        - AWSCognitoCanvaraPrototype: []
      parameters:
        - schema:
            type: integer
            default: 20
          in: query
          name: limit
          description: Maximum number of activities to return.
        - schema:
            type: integer
          in: query
          name: start
          description: Start index from where to return the activities.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  activities:
                    type: array
                    items:
                      type: object
                      properties:
                        activity_id:
                          type: string
                          description: Unique identifier of this activity.
                        type:
                          type: string
                          description: Type of activity.
                        data:
                          type: object
                          description: Data to display in the activity.
                        created_at:
                          type: string
                          description: 'UTC date and time when this activity was created, in ISO 8601 format.'
  /myactivities:
    get:
      summary: List my activities
      operationId: get_activities
      description: ''
      tags:
        - activities
      deprecated: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: object
                    description: List of all posts.
                    properties:
                      gigs:
                        type: array
                        items:
                          $ref: '#/components/schemas/Post'
                        description: List of gigs.
                      offers:
                        type: array
                        items:
                          $ref: '#/components/schemas/Offer'
                        description: List of offers.
                      positions:
                        type: array
                        items:
                          $ref: '#/components/schemas/Position'
                        description: List of positions.
                  applications:
                    type: object
                    description: List of applications.
                    properties:
                      applications:
                        type: array
                        items:
                          $ref: '#/components/schemas/Application'
                        description: List of application.
                      proposals:
                        type: array
                        items:
                          $ref: '#/components/schemas/Proposal'
                        description: List of proposals.
                  tasks:
                    type: object
                    description: List of tasks.
                    properties:
                      applications:
                        type: array
                        items:
                          $ref: '#/components/schemas/Application'
                        description: List of applications.
                      proposals:
                        type: array
                        items:
                          $ref: '#/components/schemas/Proposal'
                        description: List of proposals.
                  connections:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                    description: List of connections.
                  communities:
                    type: array
                    items:
                      $ref: '#/components/schemas/Community'
                    description: List of communities.
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                    description: List of events.
                  bookmarks:
                    type: object
                    description: List of all bookmarks.
                    properties:
                      gigs:
                        type: array
                        items:
                          $ref: '#/components/schemas/Post'
                        description: List of gigs.
                      offers:
                        type: array
                        items:
                          $ref: '#/components/schemas/OfferMinimal'
                        description: List of offers.
                      positions:
                        type: array
                        items:
                          $ref: '#/components/schemas/Position'
                        description: List of positions.
                      communities:
                        type: array
                        items:
                          $ref: '#/components/schemas/Community'
                        description: List of communities.
                      events:
                        type: array
                        items:
                          $ref: '#/components/schemas/Event'
                        description: List of events.
                      peoples:
                        type: array
                        items:
                          $ref: '#/components/schemas/UserSummary'
                        description: List of peoples.
      security:
        - AWSCognitoCanvaraPrototype: []
...
components:
  schemas:
...
    Position:
      title: Position details
      type: object
      properties:
        position_id:
          type: string
          description: Unique identifier for position.
        hiring_manager:
          type: object
          description: Information about the manager.
          required:
            - user_id
            - name
          properties:
            user_id:
              type: string
              description: Unique identifier for the user.
            name:
              type: string
              description: Name of the user.
            profile_picture_url:
              type: string
              description: The URL of user's profile picture.
            location:
              type: string
              description: Location of the higring manager.
            email:
              type: string
              description: Email of the user.
            phone_number:
              type: string
              description: Phone of the user.
        location:
          $ref: '#/components/schemas/Location'
        role_type:
          type: string
          description: Role Type of this position.
          enum:
            - Full Time
            - Part Time
        role:
          type: string
          description: Role of this position.
        department:
          type: string
          description: Department for the position.
        description:
          type: string
          description: Description of this position.
        benefits:
          type: string
          description: Benefits for this position.
        pay_currency:
          type: string
          description: Payout currency for this position.
        pay_minimum:
          type: number
          multipleOf: 0.01
          description: Minimum payout for this position.
        pay_maximum:
          type: number
          multipleOf: 0.01
          description: Maximum payout for this position.
        apply_url:
          type: string
          description: Apply URL for this position.
      required:
        - hiring_manager
        - location
        - role_type
        - role
        - description
        - pay_currency
        - pay_minimum
        - pay_maximum
    Community:
      title: Community
      type: object
      properties:
        community_id:
          type: string
          description: Unique identifier for this community.
        name:
          type: string
          description: Name of this community.
        primary_moderator:
          type: object
          description: Information about the owner/primary moderator of the community.
          required:
            - user_id
            - name
          properties:
            user_id:
              type: string
              description: Unique identifier for the user.
            name:
              type: string
              description: Name of the user.
            profile_picture_url:
              type: string
              description: The URL of user's profile picture.
        secondary_moderator:
          type: object
          description: Information about the secondary moderator of the community.
          required:
            - user_id
            - name
          properties:
            user_id:
              type: string
              description: Unique identifier for the user.
            name:
              type: string
              description: Name of the user.
            profile_picture_url:
              type: string
              description: The URL of user's profile picture.
        location:
          $ref: '#/components/schemas/Location'
        sponsor_events:
          type: array
          description: Events sponsored by this community.
          items:
            anyOf:
              - $ref: '#/components/schemas/EventSummary'
              - type: string
        community_logo:
          $ref: '#/components/schemas/UserUpload'
          description: Logo of the community.
        overview_video:
          $ref: '#/components/schemas/UserUpload'
          description: Overview video of the community.
        type:
          type: string
          description: Types of this community.
        membership_approval_required:
          type: boolean
          description: Membership approval is required or not for the community.
        mission:
          type: string
          description: Mission of this community.
        target_audience:
          type: string
          description: Target audiences of this community.
        activities:
          type: string
          description: Activities of this community.
        announcements:
          type: array
          description: Announcements of this community.
          items:
            anyOf:
              - $ref: '#/components/schemas/CommunityAnnouncement'
              - type: string
        members:
          type: array
          description: Members of this community.
          items:
            anyOf:
              - $ref: '#/components/schemas/CommunityMember'
              - type: string
        gallery:
          type: array
          items:
            anyOf:
              - $ref: '#/components/schemas/UserUpload'
              - type: string
        hashtags:
          type: array
          description: Hashtags of this community.
          items:
            type: string
        contact_email:
          type: string
          description: Email of this community.
        contact_phone:
          type: string
          description: Phone number of this community.
        contact_messaging:
          type: string
          description: Messaging id of this community.
        member_can_create_events:
          type: boolean
          description: Members are allowed to create events or not.
        member_can_post_photos:
          type: boolean
          description: Members are allowed to post photos or not.
        member_can_post_updates:
          type: boolean
          description: Members are allowed to post updates or not.
        created_at:
          type: string
          description: 'Date and time when this community was created, in ISO 8601 format.'
        last_updated_at:
          type: string
          description: 'Date and time when this community was last updated, in ISO 8601 format.'
      required:
        - name
        - location_id
        - type
        - membership_approval_required
        - mission
...
    User:
      title: User
      type: object
      properties:
        customer_id:
          type: string
        user_id:
          type: string
        username:
          type: string
        name:
          type: string
        title:
          type: string
        profile_picture_url:
          type: string
        background_picture_url:
          type: string
        location:
          type: string
        customer_name:
          type: string
        email:
          type: string
        phone_number:
          type: string
        linkedin_url:
          type: string
        manager:
          type: object
          properties:
            user_id:
              type: string
            name:
              type: string
        product_preferences:
          type: array
          items:
            $ref: '#/components/schemas/ProductPreference'
        current_skills:
          type: array
          items:
            $ref: '#/components/schemas/SkillWithLevelAndExpert'
        desired_skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        fun_facts:
          type: array
          items:
            anyOf:
              - $ref: '#/components/schemas/UserUpload'
              - type: string
        introduction:
          type: string
        introduction_video:
          $ref: '#/components/schemas/UserUpload'
        career_goals:
          type: string
        career_goals_visibility:
          type: boolean
        hidden_secrets:
          type: string
        superpowers:
          type: string
        trying_to_learn:
          type: string
        trying_to_learn_visibility:
          type: boolean
        company_start_date:
          type: string
        pronoun:
          type: string
        department:
          type: string
        languages:
          type: array
          items:
            type: string
        feedback:
          type: array
          items:
            $ref: '#/components/schemas/Feedback'
        allow_demo_mode:
          type: boolean
        onboarding_complete:
          type: boolean
        hashtags:
          type: array
          items:
            type: string
        slack_teams_messaging_id:
          type: string
        mentorship_offered:
          type: boolean
        mentorship_description:
          type: string
        mentorship_hashtags:
          type: array
          items:
            type: string
        mentorship_video:
          $ref: '#/components/schemas/UserUpload'
        resume_file:
          $ref: '#/components/schemas/UserUpload'
        community_memberships:
          type: array
          description: Community memberships for user.
          items:
            anyOf:
              - $ref: '#/components/schemas/CommunitySummary'
              - type: string
        relevant_skills_recommendation:
          type: array
          items:
            type: string
        linkedin_profiles_suggestion:
          type: array
          items:
            type: string
      required:
        - customer_id
        - user_id
        - name
...
    Offer:
      allOf:
        - $ref: '#/components/schemas/OfferMinimal'
        - title: Offer
        - type: object
          properties:
            overview_video:
              $ref: '#/components/schemas/UserUpload'
              description: Overview video for the offer.
...
    OfferMinimal:
      title: Offer
      type: object
      properties:
        offer_id:
          type: string
          description: Unique identifier for this offer.
        offerer:
          type: object
          description: Information about the owner/offerer of the offer.
          properties:
            user_id:
              type: string
              description: Unique identifier for the user.
            name:
              type: string
              description: Name of the user.
            profile_picture_url:
              type: string
              description: The URL of user's profile picture.
            title:
              type: string
              description: Title of the user.
            pronoun:
              type: string
              description: Pronoun.
            location:
              type: string
              description: Location of the user.
            department:
              type: string
              description: Department of the user.
            email:
              type: string
              description: Email of the user.
            phone_number:
              type: string
              description: Phone of the user.
        name:
          type: string
          description: Name of this offer.
        overview_text:
          type: string
          description: Overview text for the offer.
        hashtags:
          type: array
          description: Hashtags of this offer.
          items:
            type: string
        status:
          type: string
          description: Current status of this offer.
        created_at:
          type: string
          description: 'Date and time when this offer was created, in ISO 8601 format.'
        last_updated_at:
          type: string
          description: 'Date and time when this offer was last updated, in ISO 8601 format.'
        proposal_count:
          type: number
          description: Number of proposals belongs to this offer
      required:
        - name
        - overview_text
    Proposal:
      allOf:
        - $ref: '#/components/schemas/ProposalMinimal'
        - title: Proposal
        - type: object
          properties:
            overview_video:
              $ref: '#/components/schemas/UserUpload'
              description: Overview video for the proposal.
...
    Post:
      title: Post
      type: object
      description: ''
      properties:
        customer_id:
          type: string
          deprecated: true
          description: Unique identifier of the customer to which this post belongs.
        post_id:
          type: string
          description: Unique identifier for this post.
        post_owner_id:
          type: string
          deprecated: true
          description: Unique identifier of the owner of this post.
        post_owner:
          type: object
          description: Information about the owner of this post.
          required:
            - user_id
            - name
          properties:
            user_id:
              type: string
              description: Unique identifier for the user.
            name:
              type: string
              description: Name of the user.
            profile_picture_url:
              type: string
              description: The URL of user's profile picture.
            location:
              type: string
              description: Location of the post owner.
        task_owner_id:
          type: string
          deprecated: true
        post_type:
          $ref: '#/components/schemas/PostType'
        name:
          type: string
          description: Name of this post.
        status:
          type: string
          description: Current status of this post.
        summary:
          type: string
          deprecated: true
          description: Name of this post.
        description:
          type: string
          description: Description of this post.
        video_url:
          type: string
          description: The URL of video description for this post.
        description_video:
          $ref: '#/components/schemas/UserUpload'
        size:
          type: string
          description: T-shirt size of this post.
          enum:
            - S
            - M
            - L
        location:
          $ref: '#/components/schemas/Location'
        people_needed:
          type: integer
          description: Number of people needed for this post.
        language:
          type: string
          description: Language for this post.
        required_skills:
          type: array
          description: Must-have skills for this post.
          items:
            $ref: '#/components/schemas/SkillWithLevel'
        desired_skills:
          type: array
          description: Nice-to-have skills for this post.
          items:
            $ref: '#/components/schemas/SkillWithLevel'
        candidate_description:
          type: string
          description: Description of an ideal candidate for this post.
        auto_assign:
          type: boolean
          description: Automatically select applications or not.
        canvara_kudos:
          type: boolean
          description: Canvara kudos true or false.
        hashtags:
          type: array
          items:
            type: string
          description: List of hastags or keywords.
        highlighted_communities:
          type: array
          description: Highlighted Community ids for gig.
          items:
            type: object
            description: Information about the community.
            required:
              - community_id
              - name
            properties:
              community_id:
                type: string
                description: Unique identifier for the community.
              name:
                type: string
                description: Name of the community.
              community_logo:
                type: string
                description: The URL of community logo.
        match_level:
          type: number
          description: A number between 0 and 100 indicating how close of a match this post is for the current user. Returned only for active posts and if the current user is not the post owner.
        matched_skills:
          type: array
          description: Post's skills that match the user's skills.
          items:
            $ref: '#/components/schemas/SkillWithLevel'
        unmatched_skills:
          type: array
          description: Post's skills that don't match the user's skills.
          items:
            $ref: '#/components/schemas/SkillWithLevel'
        is_bookmarked:
          type: boolean
          description: Whether this post has been bookmarked by the current user or not.
        is_liked:
          type: boolean
          description: Whether this post has been liked by the current user or not.
        target_date:
          type: string
          description: 'Target date for the completion of work described in this post, in ISO 8601 format.'
        expiration_date:
          type: string
          description: ISO 8601 formatted Date on which this post expires if no candidate is selected.
        created_at:
          type: string
          description: 'Date and time when this post was created, in ISO 8601 format.'
        last_updated_at:
          type: string
          description: 'Date and time when this post was last updated, in ISO 8601 format.'
      required:
        - post_id
        - post_owner
        - post_type
        - name
        - status
        - size
        - location
        - people_needed
        - language
        - target_date
...
    Application:
      title: Application
      type: object
      properties:
        application_id:
          type: string
          description: Unique identifier for this application.
        post_id:
          type: string
          description: Unique identifier for the post.
        post:
          type: object
          description: Information about post.
          properties:
            post_id:
              type: string
              description: Unique identifier for this post.
            owner:
              type: object
              description: Information about the post owner.
              properties:
                user_id:
                  type: string
                  description: Unique identifier for the user.
                name:
                  type: string
                  description: Name of the user.
                profile_picture_url:
                  type: string
                  description: The URL of user's profile picture.
                title:
                  type: string
                  description: Title of the user.
                pronoun:
                  type: string
                  description: Pronoun.
                location:
                  type: string
                  description: Location of the user.
                department:
                  type: string
                  description: Department of the user.
            name:
              type: string
              description: Name of the post.
            description:
              type: string
              description: Description of this post.
            hashtags:
              type: array
              items:
                type: string
              description: List of hastags or keywords.
            status:
              type: string
              description: Current status of this post.
            created_at:
              type: string
              description: 'Date and time when this post was created, in ISO 8601 format.'
            last_updated_at:
              type: string
              description: 'Date and time when this post was last updated, in ISO 8601 format.'
        applicant_id:
          type: string
          description: Unique identifier for the user applying for the post.
        applicant:
          type: object
          description: Information about the applicant.
          properties:
            user_id:
              type: string
              description: Unique identifier for the user.
            name:
              type: string
              description: Name of the user.
            profile_picture_url:
              type: string
              description: The URL of user's profile picture.
            title:
              type: string
              description: Title of the user.
            pronoun:
              type: string
              description: Pronoun.
            location:
              type: string
              description: Location of the user.
            department:
              type: string
              description: Department of the user.
            email:
              type: string
              description: Email of the user.
            phone_number:
              type: string
              description: Phone number of the user.
            slack_teams_messaging_id:
              type: string
              description: Messaging handle of the user.
        description:
          type: string
          description: Additional information for this application.
        video_url:
          type: string
          description: The URL of video for this application.
        description_video:
          $ref: '#/components/schemas/UserUpload'
        matched_skills:
          type: array
          description: Post's skills that match the user's skills.
          items:
            $ref: '#/components/schemas/SkillWithLevel'
        unmatched_skills:
          type: array
          description: Post's skills that don't match the user's skills.
          items:
            $ref: '#/components/schemas/SkillWithLevel'
        status:
          type: string
          description: Current status of this application.
        created_at:
          type: string
          description: 'Date and time when this application was created, in ISO 8601 format.'
        last_updated_at:
          type: string
          description: 'Date and time when this application was last updated, in ISO 8601 format.'
...
    Skill:
      title: Skill
      type: object
      properties:
        skill_id:
          type: string
        name:
          type: string
...
  securitySchemes:
    AWSCognitoCanvaraPrototype:
      type: http
      scheme: bearer